+++ tests/php/screen/TestSettings.php
@@ -198,6 +198,51 @@ class TestSettings extends BaseTestCase {
 	}
 
 	/**
+	 * Test that checking the "remove token" checkbox clears the stored token.
+	 *
+	 * @group screen
+	 * @group settings-screen
+	 */
+	public function test_action_admin_init_remove_token_checkbox_clears_stored_token() {
+		global $_POST;
+
+		if ( defined( 'EP_CREDENTIALS' ) && EP_CREDENTIALS ) {
+			$this->markTestSkipped( 'EP_CREDENTIALS constant overrides the option.' );
+		}
+
+		// Make is_epio() true so get_epio_credentials() reads the option.
+		putenv( 'IS_EPIO_ENVIRONMENT=1' ); // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.runtime_configuration_putenv
+
+		$settings = new Settings();
+
+		Utils\update_option(
+			'ep_credentials',
+			[
+				'username' => 'u',
+				'token'    => 'secret',
+			]
+		);
+
+		$_POST = [
+			'ep_settings_nonce' => wp_create_nonce( 'elasticpress_settings' ),
+			'ep_language'       => 'site-default',
+			'ep_host'           => Utils\get_host(),
+			'ep_credentials'    => [
+				'username' => 'u',
+				'token'    => '',
+			],
+			'ep_remove_token'   => '1',
+		];
+
+		$settings->action_admin_init();
+
+		$this->assertSame( '', Utils\get_option( 'ep_credentials' )['token'] );
+		$this->assertSame( 'u', Utils\get_option( 'ep_credentials' )['username'] );
+
+		putenv( 'IS_EPIO_ENVIRONMENT' ); // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.runtime_configuration_putenv
+	}
+
+	/**
 	 * Test the `add_validation_notice` method
 	 *
 	 * @group screen